requirements.txt
fastapi
lightgbm
pandas
numpy
joblib
python-dateutil